Universe and Wars Added
Added both APIs.

To:Do test the POSTS, first time encountering it, slight refactoring needed but need to get some correct IDs and use cases to really test it and try it. All gets appear to be working but dummy data is needed, the end points are being hit correctly though